+++ packages/sit-onyx/src/components/OnyxCheckbox/OnyxCheckbox.ct.tsx
@@ -1,5 +1,5 @@
 import { DENSITIES } from "../../composables/density";
-import { expect, test } from "../../playwright-axe";
+import { expect, test } from "../../playwright/a11y";
 import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
 import { TRUNCATION_TYPES } from "../../types";
 import OnyxCheckbox from "./OnyxCheckbox.vue";
