+++ packages/sit-onyx/src/components/OnyxCheckboxGroup/OnyxCheckboxGroup.ct.tsx
@@ -1,4 +1,4 @@
-import { expect, test } from "../../playwright-axe";
+import { expect, test } from "../../playwright/a11y";
 import { DIRECTIONS } from "../../types";
 import OnyxCheckboxGroup from "./OnyxCheckboxGroup.vue";
 import type { OnyxCheckboxGroupProps } from "./types";
