+++ packages/sit-onyx/src/components/OnyxEmpty/OnyxEmpty.ct.tsx
@@ -1,4 +1,4 @@
-import { expect, test } from "../../playwright-axe";
+import { expect, test } from "../../playwright/a11y";
 import { mockPlaywrightIcon } from "../../playwright/screenshots";
 import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
 import OnyxEmpty from "./OnyxEmpty.vue";
