+++ packages/sit-onyx/src/components/OnyxIconButton/OnyxIconButton.ct.tsx
@@ -1,5 +1,5 @@
 import { DENSITIES } from "../../composables/density";
-import { expect, test } from "../../playwright-axe";
+import { expect, test } from "../../playwright/a11y";
 import {
   executeMatrixScreenshotTest,
   mockPlaywrightIcon,
