+++ packages/sit-onyx/src/components/OnyxLink/OnyxLink.ct.tsx
@@ -1,4 +1,4 @@
-import { expect, test } from "../../playwright-axe";
+import { expect, test } from "../../playwright/a11y";
 import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
 import OnyxLink from "./OnyxLink.vue";
