+++ packages/sit-onyx/src/components/OnyxListbox/OnyxListbox.ct.tsx
@@ -1,5 +1,5 @@
 import { listboxTesting } from "@sit-onyx/headless/playwright";
-import { expect, test } from "../../playwright-axe";
+import { expect, test } from "../../playwright/a11y";
 import OnyxButton from "../OnyxButton/OnyxButton.vue";
 import OnyxListbox from "./OnyxListbox.vue";
 import type { ListboxOption, OnyxListboxProps } from "./types";
