+++ packages/sit-onyx/src/components/OnyxLoadingIndicator/OnyxLoadingIndicator.ct.tsx
@@ -1,4 +1,4 @@
-import { test } from "../../playwright-axe";
+import { test } from "../../playwright/a11y";
 import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
 import OnyxLoadingIndicator from "./OnyxLoadingIndicator.vue";
