+++ packages/sit-onyx/src/components/OnyxSelect/OnyxSelect.ct.tsx
@@ -1,5 +1,5 @@
 import { DENSITIES } from "../../composables/density";
-import { expect, test } from "../../playwright-axe";
+import { expect, test } from "../../playwright/a11y";
 import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
 import OnyxSelect from "./OnyxSelect.vue";
 import { MULTISELECT_TEXT_MODE } from "./types";
