+++ packages/sit-onyx/src/components/OnyxTooltip/OnyxTooltip.ct.tsx
@@ -1,4 +1,4 @@
-import { expect, test } from "../../playwright-axe";
+import { expect, test } from "../../playwright/a11y";
 import { executeMatrixScreenshotTest, mockPlaywrightIcon } from "../../playwright/screenshots";
 import OnyxTooltip from "./OnyxTooltip.vue";
